LayoutTests/imported/w3c/web-platform-tests/editing/other/merge-span-with-style-after-backspace-having-contenteditable-expected.txt

Back

pace


FAIL waiting for command to execute assert_equals: Style is not preserved for the span after pressing backspace in contenteditable expected "<h1><span style=\"background-color:red;\">Back</span><span style=\"background-color: red;\">space</span></h1>" but got "<h1><span style=\"background-color:red;\">Back</span></h1>\n<h1><span style=\"background-color: red;\" id=\"space\">pace</span></h1>"
